Relocate components used in multi-profile
