+++ src/features/multi-profile/components/ProfileCard.tsx
@@ -4,12 +4,13 @@ import { Card } from 'native-base';
 
 import { AvatarName, getAvatarByName } from '@covid/utils/avatar';
 import { getDaysAgo } from '@covid/utils/datetime';
-import { GreenTick } from '@covid/components/GreenTick';
 import { colors } from '@theme';
 import i18n from '@covid/locale/i18n';
 import { ClippedText, SecondaryText } from '@covid/components/Text';
-import LastReported from '@covid/components/LastReported';
 import { Profile } from '@covid/core/profile/ProfileService';
+
+import LastReported from './LastReported';
+import { GreenTick } from './GreenTick';
 
 type Props = {
   profile: Profile;
